requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-telegram-bot==21.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
jinja2==3.1.5
python-multipart==0.0.9
itsdangerous==2.2.0
requests==2.32.3